First steps in calculating dynamic section
I want to have the clocks main hand advance with constant speed. For
this, I need to resize the 12 sections that make up the clock
dynamically. This is because the duration of these 12 sections is
connected to the sunrise and sunset. So the 12 sections are different
every day.

Here, I added this constantly advancing main hand, together with two
others that mark the sunrise and sunset on the clock.

I rebuilt the whole logic for getting the sunset and sunrise. I want to
display the sunset and sunrise that is closest to the current time. Not
the sunset for the current day. This probably only makes a small
difference, but it feels right to do it this way.